# Optimize message formatting