@OsirisTerje 2024-08-19
> On an related note, once net462 is dropped, all the tuple checks can be drastically simplified by using the [ITuple](https://learn.microsoft.com/en-us/dotnet/api/system.runtime.compilerservices.ituple) interface.

Any reason not to do this now, using compile conditionals ?