@RenderMichael 2024-08-19
@OsirisTerje Not as far as I’m aware, but that’s a more in-depth refactoring that affects more than just logging. I noticed this quick improvement I can make. 

Also, checking against `ITuple` directly means that anyone who implements that type themselves (which is probably nobody) would see a change in output.